refactor: move edition-agnostic request handling to the platform
First of a stack teasing the platform apart from the n23 edition. This one
touches no models and no database — only code with zero edition
dependencies, verified module by module.

Moved to gyrinx/:
- adapter.py — the allauth adapter, which imported only allauth, django
  and gyrinx.admin_site
- impersonation.py — session-key helpers with no imports at all
- middleware.py — ClearLoggingRequestMiddleware and
  RequestSizeExceptionMiddleware (its errors/error.html template was
  already platform-side)
- context_processors.py — gyrinx_debug and impersonation, which read only
  settings and request attributes
- account_forms.py — the three allauth overrides named in ACCOUNT_FORMS,
  plus UsernameChangeForm
- widgets.py — the four Bootstrap form widgets, whose templates already
  lived in gyrinx/pages/templates/

Left behind deliberately, each blocked on a model that is still edition:
ImpersonationMiddleware (ImpersonationLog), site_banner (Banner),
notifications (Notification), BadgeSelectionForm (UserProfile).

Also breaks the gyrinx.forms <-> n23 cycle. Both offending helpers were
duck-typed — fighter_group_key reads fighter.house.name,
template_form_with_terms reads fighter.term_* — so the annotations were
the only hard link. They move to n23/core/forms/terms.py, where they
belong semantically; the generic group_select and group_sorter stay.
Nothing in the platform imported gyrinx.forms at all.

test_request_data_too_big.py follows the middleware into gyrinx/tests/,
and the bandit baseline is regenerated because it keys on file path.

Platform->edition references drop from 34 to 23. 4166 tests pass;
makemigrations --check reports no changes.

Refs #2093
