+++ gyrinx/adapter.py
