+++ gyrinx/forms.py
@@ -1,11 +1,5 @@
 from collections.abc import Callable
 from itertools import groupby
-from typing import Optional
-
-from django import forms
-
-from n23.content.models import ContentFighter
-from n23.core.models.list import ListFighter
 
 
 def group_select(form, field, key=lambda x: x, sort_groups_by=None):
@@ -54,11 +48,6 @@ def group_select(form, field, key=lambda x: x, sort_groups_by=None):
         formfield.widget.choices = choices
 
 
-def fighter_group_key(fighter: ContentFighter):
-    # Group by house name
-    return fighter.house.name
-
-
 def group_sorter(priority_name: str) -> Callable[[str], tuple]:
     def sort_groups_key(group_name):
         # Sort groups: gang's own house first, then alphabetically
@@ -68,26 +57,3 @@ def group_sorter(priority_name: str) -> Callable[[str], tuple]:
             return (1, group_name)  # Other houses alphabetically
 
     return sort_groups_key
-
-
-def template_form_with_terms(form: forms.Form, fighter: Optional[ListFighter] = None):
-    # Get the correct terminology for this fighter
-    terms = dict(
-        term_singular=fighter.term_singular if fighter else "Fighter",
-        term_singular__lower=fighter.term_singular.lower() if fighter else "fighter",
-        term_injury_singular=fighter.term_injury_singular if fighter else "Injury",
-        term_injury_singular__lower=fighter.term_injury_singular.lower()
-        if fighter
-        else "injury",
-        term_proximal_demonstrative__lower=fighter.term_proximal_demonstrative.lower()
-        if fighter
-        else "this fighter",
-        term_proximal_demonstrative=fighter.term_proximal_demonstrative
-        if fighter
-        else "This fighter",
-    )
-    for field in form.fields.values():
-        if hasattr(field, "help_text"):
-            field.help_text = field.help_text.format(**terms)
-        if hasattr(field, "label"):
-            field.label = field.label.format(**terms)
