+++ n23/core/context_processors.py
@@ -1,6 +1,5 @@
 import logging
 
-from django.conf import settings
 from django.core.cache import cache
 from django.db import DatabaseError, InterfaceError, OperationalError
 
@@ -63,24 +62,6 @@ def site_banner(request):
             context["banner"] = live_banner
 
     return context
-
-
-def gyrinx_debug(request):
-    """Add gyrinx_debug flag to the context for debug UI elements."""
-    return {"gyrinx_debug": settings.GYRINX_DEBUG}
-
-
-def impersonation(request):
-    """Expose impersonation state to templates.
-
-    Set by :class:`n23.core.middleware.ImpersonationMiddleware`. When
-    ``is_impersonating`` is true, ``request.user`` is the impersonated user and
-    ``impersonator`` is the real admin.
-    """
-    return {
-        "is_impersonating": getattr(request, "is_impersonating", False),
-        "impersonator": getattr(request, "impersonator", None),
-    }
 
 
 def notifications(request):
