+++ n23/core/forms/vehicle.py
@@ -12,7 +12,8 @@ from n23.content.models import (
 )
 from n23.core.forms.list import ContentFighterChoiceField, group_select
 from n23.core.models.list import List
-from gyrinx.forms import fighter_group_key, group_sorter
+from n23.core.forms.terms import fighter_group_key
+from gyrinx.forms import group_sorter
 from gyrinx.models import FighterCategoryChoices
